requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.0

# LangChain ecosystem
langchain==0.2.11
langchain-core==0.2.24
langchain-community==0.2.10
langchain-qdrant==0.1.3
langchain-openai==0.1.17
langchain-groq==0.1.9
langsmith==0.1.95

# Vector DB
qdrant-client==1.10.1

# Document parsing
docling==2.5.0

# Auth
python-jose[cryptography]==3.3.0
cryptography==42.0.8

# Guardrails - PII
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.5

# Evaluation
ragas==0.1.21
datasets==2.20.0

# Azure observability
azure-monitor-opentelemetry==1.6.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0

# Utilities
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
structlog==24.2.0
tenacity==8.4.1
